Updated Projects to .NET Core 3.0 Updated all projects that had some form of ASP.NET dependency. They will now work with new core package reference. They also became TargetFramework of netcoreapp3.0. With a FrameworkReference of Microsoft.AspNetCore.App. Fixed Unit Tests that broke. Mostly removed logger checking, to fragile. Fixed ServerCore registration and status checking. Created check, disconnected, and registration handlers for the  connection. Updated logging, will now show hosting lifetime events.
In-depth analysis can be found in upgrade-dotnet-core-3.0.todo.

+++ test/EventHorizon.Zone.Core.Tests/Json/NewtonsoftJsonFileSaverTests.cs
@@ -18,20 +18,25 @@ namespace EventHorizon.Zone.Core.Tests.Json
         public NewtonsoftJsonFileSaverTests()
         {
             // Remove existing test file
-            var fullFileName = Path.Combine(
+            var fileFullName = Path.Combine(
                 directory,
                 fileName
             );
-            File.Delete(
-                fullFileName
-            );
+            if (File.Exists(
+                fileFullName
+            ))
+            {
+                File.Delete(
+                    fileFullName
+                );
+            }
         }
 
         [Fact]
         public async Task TestShouldSaveSerializedRepresentationWhenObjectIsPassedToSpecifiedDirectoryAndFileName()
         {
             // Given 
-            var fullFileName = Path.Combine(
+            var fileFullName = Path.Combine(
                 directory,
                 fileName
             );
@@ -43,7 +48,7 @@ namespace EventHorizon.Zone.Core.Tests.Json
             // Verify File is not there
             Assert.False(
                 File.Exists(
-                    fullFileName
+                    fileFullName
                 )
             );
 
@@ -59,7 +64,7 @@ namespace EventHorizon.Zone.Core.Tests.Json
             // Then
             Assert.True(
                 File.Exists(
-                    fullFileName
+                    fileFullName
                 )
             );
         }
